restructure home components
